chore: move mid-price ratio to 'prices' module
