+++ src/app/prices/index.ts
@@ -1,3 +1,4 @@
 export * from "./get-current-price"
 export * from "./get-price-history"
 export * from "./list-currencies"
+export * from "./mid-price"
