+++ src/app/prices/mid-price.ts
