--- src/app/shared/index.ts
@@ -1,1 +1,0 @@
-export * from "./mid-price"
