+++ test/integration/02-user-wallet/02-receive-onchain.spec.ts
@@ -2,7 +2,7 @@
 import { once } from "events"
 
 import { Accounts, Prices, Wallets } from "@app"
-import { usdFromBtcMidPriceFn } from "@app/shared"
+import { usdFromBtcMidPriceFn } from "@app/prices"
 
 import {
   getAccountLimits,
