+++ test/integration/02-user-wallet/02-send-lightning-limits.spec.ts
@@ -1,6 +1,6 @@
 /* eslint jest/expect-expect: ["error", { "assertFunctionNames": ["expect", "successLimitsPaymentTests", "multiplePaymentLimitTests"] }] */
 import { Payments, Wallets } from "@app"
-import { getMidPriceRatio } from "@app/shared"
+import { getMidPriceRatio } from "@app/prices"
 
 import { getDealerConfig } from "@config"
